This push fixes the issue of displaying the stock status for each drug. The status is now linked to the current drug model qty
