+++ drugs/models.py
@@ -42,7 +42,7 @@ class Drug(models.Model):
     side_effects = models.TextField(blank=True, null=True) # Common side effects.
     contraind    = models.TextField(blank=True, null=True) # Conditions that serve as reasons to not use the drug.
     dosage_inst  = models.TextField(blank=True, null=True) # Recommended dosage.
-    stock_qty    = models.PositiveIntegerField() # Number of units available.
+    stock_qty    = models.PositiveIntegerField(default=0) # Number of units available.
     exp_date     = models.DateField() # Expiration date.
     rx_required  = models.CharField(max_length=20, choices=RX_CHOICE) # Is a prescription is necessary.
     price        = models.DecimalField(max_digits=10, decimal_places=2) # Cost of the drug per unit.
@@ -53,3 +53,11 @@ class Drug(models.Model):
 
     def __str__(self):
         return self.drug_name
+    
+    def stock_status(self):
+        if self.stock_qty <= 0:
+            return "Out of Stock"
+        elif self.stock_qty < 120:
+            return "Low Stock"
+        else:
+            return "In Stock"
